Add support for `authorize` flow
This adds support for the `authorize` flow in test app.
The test page app was refactored in other to support
multiple flows easier. Also, configs are now centralized
under `config` folder instead of using environment variables.

Typings of the `PaymentWidget` class were also changed
to reflect the breaking change for the `complete` event
and now it accepts a type parameter that allows to specify
the expected flow returning value for the `complete` event.