Add support for `authorize` flow
This adds support for the `authorize` flow in test app
by using the env var `VITE_PAYMENT_WIDGET_FLOW`
to specify the desired flow of the widget.

Typings of the `PaymentWidget` class were also changed
to reflect the breaking change for the `complete` event
and now it accepts a type parameter that allows to specify
the expected flow returning value for the `complete` event.

Finally, proxying requests were re-added as it might be
useful to test locally and avoid CORS errors.